Moved classes and interfaces to appropriate namespaces
